+++ src-tauri/src/os/mod.rs
@@ -1,1 +1,2 @@
+pub mod hotkeys;
 pub mod tray;